storage moved, and an interface proposal
