+++ client/src/app.tsx
@@ -4,7 +4,6 @@ import {BehaviorSubject} from "rxjs/BehaviorSubject"
 import {Subscription} from "rxjs/Subscription"
 import {PerspectiveCamera} from "three"
 
-import {AppStorage} from "./app-storage"
 import {IFabricExports} from "./body/fabric-exports"
 import {createFabricKernel, FabricKernel} from "./body/fabric-kernel"
 import {Physics} from "./body/physics"
@@ -12,6 +11,7 @@ import {INITIAL_JOINT_COUNT, MAX_POPULATION} from "./gotchi/evolution"
 import {Island} from "./island/island"
 import {IslandState} from "./island/island-state"
 import {Surface} from "./island/spot"
+import {AppStorage} from "./storage/app-storage"
 import {ActionsPanel} from "./view/actions-panel"
 import {GotchiView} from "./view/gotchi-view"
 import {InfoPanel} from "./view/info-panel"
