+++ client/src/index.tsx
@@ -3,10 +3,11 @@ import * as React from "react"
 import * as ReactDOM from "react-dom"
 
 import App from "./app"
-import {AppStorage} from "./app-storage"
 import {IFabricExports} from "./body/fabric-exports"
-import "./index.css"
 import registerServiceWorker from "./service-worker"
+import {AppStorage} from "./storage/app-storage"
+
+import "./index.css"
 
 declare const getFabricExports: () => Promise<IFabricExports> // implementation: index.html
 const storage: AppStorage = new AppStorage(localStorage)
